refactor: rename Step Functions CloudWatch error logging rule files
