+++ src/rules/stepfunctions/index.ts
@@ -2,5 +2,5 @@
 Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
-export { default as StepFunctionStateMachineAllLogsToCloudWatch } from './StepFunctionStateMachineAllLogsToCloudWatch';
+export { default as StepFunctionStateMachineErrorLogsToCloudWatch } from './StepFunctionStateMachineErrorLogsToCloudWatch';
 export { default as StepFunctionStateMachineXray } from './StepFunctionStateMachineXray';
